+++ pkg/provider/llm_blueprint_resource.go
@@ -6,12 +6,22 @@ import (
 	"fmt"
 
 	"github.com/datarobot-community/terraform-provider-datarobot/internal/client"
+	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/objectvalidator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/objectplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+const (
+	defaultPromptType = "CHAT_HISTORY_AWARE"
 )
 
 // Ensure provider defined types fully satisfy framework interfaces.
@@ -62,7 +72,7 @@ func (r *LLMBlueprintResource) Schema(ctx context.Context, req resource.SchemaRe
 				MarkdownDescription: "The id of the Vector Database for the LLM Blueprint.",
 				Optional:            true,
 				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplaceIfConfigured(),
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
 			"llm_id": schema.StringAttribute{
@@ -71,6 +81,66 @@ func (r *LLMBlueprintResource) Schema(ctx context.Context, req resource.SchemaRe
 				PlanModifiers: []planmodifier.String{
 					// in order to generate an update to the custom model resource, we need to force a replace
 					stringplanmodifier.RequiresReplace(),
+				},
+			},
+			"llm_settings": schema.SingleNestedAttribute{
+				Optional:            true,
+				MarkdownDescription: "The LLM settings for the LLM Blueprint.",
+				PlanModifiers: []planmodifier.Object{
+					objectplanmodifier.RequiresReplace(),
+				},
+				Attributes: map[string]schema.Attribute{
+					"max_completion_length": schema.Int64Attribute{
+						MarkdownDescription: "The maximum number of tokens allowed in the completion. The combined count of this value and prompt tokens must be below the model's maximum context size, where prompt token count is comprised of system prompt, user prompt, recent chat history, and vector database citations.",
+						Optional:            true,
+					},
+					"temperature": schema.Float64Attribute{
+						MarkdownDescription: "Controls the randomness of model output, where higher values return more diverse output and lower values return more deterministic results.",
+						Optional:            true,
+					},
+					"top_p": schema.Float64Attribute{
+						MarkdownDescription: "Threshold that controls the selection of words included in the response, based on a cumulative probability cutoff for token selection. Higher numbers return more diverse options for outputs.",
+						Optional:            true,
+					},
+					"system_prompt": schema.StringAttribute{
+						MarkdownDescription: "Guides the style of the LLM response. It is a 'universal' prompt, prepended to all individual prompts.",
+						Optional:            true,
+					},
+				},
+			},
+			"prompt_type": schema.StringAttribute{
+				MarkdownDescription: "The prompt type for the LLM Blueprint.",
+				Optional:            true,
+				Computed:            true,
+				Default:             stringdefault.StaticString(defaultPromptType),
+				Validators: []validator.String{
+					stringvalidator.OneOf(defaultPromptType, "ONE_TIME_PROMPT"),
+				},
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
+			"vector_database_settings": schema.SingleNestedAttribute{
+				Optional:            true,
+				MarkdownDescription: "The Vector Database settings for the LLM Blueprint.",
+				PlanModifiers: []planmodifier.Object{
+					objectplanmodifier.RequiresReplace(),
+				},
+				Validators: []validator.Object{
+					objectvalidator.AlsoRequires(path.MatchRoot("vector_database_id")),
+				},
+				Attributes: map[string]schema.Attribute{
+					"max_documents_retrieved_per_prompt": schema.Int64Attribute{
+						MarkdownDescription: "The maximum number of documents to retrieve from the Vector Database.",
+						Optional:            true,
+						Validators: []validator.Int64{
+							int64validator.Between(0, 10),
+						},
+					},
+					"max_tokens": schema.Int64Attribute{
+						MarkdownDescription: "The maximum number of tokens to retrieve from the Vector Database.",
+						Optional:            true,
+					},
 				},
 			},
 		},
@@ -100,41 +170,12 @@ func (r *LLMBlueprintResource) Create(ctx context.Context, req resource.CreateRe
 		return
 	}
 
-	var playgroundID string
-	if IsKnown(data.PlaygroundID) {
-		playgroundID = data.PlaygroundID.ValueString()
-		_, err := r.provider.service.GetPlayground(ctx, playgroundID)
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error getting LLM Blueprint",
-				fmt.Sprintf("Unable to get LLM Blueprint, got error: %s", err),
-			)
-			return
-		}
-	}
-
-	var vectorDatabaseID string
-	if IsKnown(data.VectorDatabaseID) {
-		vectorDatabaseID = data.VectorDatabaseID.ValueString()
-		_, err := r.provider.service.GetVectorDatabase(ctx, vectorDatabaseID)
-		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error getting LLM Blueprint",
-				fmt.Sprintf("Unable to get LLM Blueprint, got error: %s", err),
-			)
-			return
-		}
-	}
-
 	var llmID string
 	if IsKnown(data.LLMID) {
 		llmID = data.LLMID.ValueString()
 		listResp, err := r.provider.service.ListLLMs(ctx)
 		if err != nil {
-			resp.Diagnostics.AddError(
-				"Error getting LLM Blueprint",
-				fmt.Sprintf("Unable to get LLM Blueprint, got error: %s", err),
-			)
+			resp.Diagnostics.AddError("Error getting LLM Blueprint", err.Error())
 			return
 		}
 
@@ -156,14 +197,43 @@ func (r *LLMBlueprintResource) Create(ctx context.Context, req resource.CreateRe
 		}
 	}
 
-	traceAPICall("CreateLLMBlueprint")
-	createResp, err := r.provider.service.CreateLLMBlueprint(ctx, &client.CreateLLMBlueprintRequest{
+	createLLMBlueprintRequest := &client.CreateLLMBlueprintRequest{
 		Name:             data.Name.ValueString(),
 		Description:      data.Description.ValueString(),
 		PlaygroundID:     data.PlaygroundID.ValueString(),
 		VectorDatabaseID: data.VectorDatabaseID.ValueString(),
 		LLMID:            data.LLMID.ValueString(),
-	})
+		PromptType:       data.PromptType.ValueString(),
+	}
+
+	if data.LLMSettings != nil {
+		createLLMBlueprintRequest.LLMSettings = &client.LLMSettings{}
+		if IsKnown(data.LLMSettings.MaxCompletionLength) {
+			createLLMBlueprintRequest.LLMSettings.MaxCompletionLength = data.LLMSettings.MaxCompletionLength.ValueInt64()
+		}
+		if IsKnown(data.LLMSettings.Temperature) {
+			createLLMBlueprintRequest.LLMSettings.Temperature = data.LLMSettings.Temperature.ValueFloat64()
+		}
+		if IsKnown(data.LLMSettings.TopP) {
+			createLLMBlueprintRequest.LLMSettings.TopP = data.LLMSettings.TopP.ValueFloat64()
+		}
+		if IsKnown(data.LLMSettings.SystemPrompt) {
+			createLLMBlueprintRequest.LLMSettings.SystemPrompt = data.LLMSettings.SystemPrompt.ValueString()
+		}
+	}
+
+	if data.VectorDatabaseSettings != nil {
+		createLLMBlueprintRequest.VectorDatabaseSettings = &client.VectorDatabaseSettings{}
+		if IsKnown(data.VectorDatabaseSettings.MaxDocumentsRetrievedPerPrompt) {
+			createLLMBlueprintRequest.VectorDatabaseSettings.MaxDocumentsRetrievedPerPrompt = data.VectorDatabaseSettings.MaxDocumentsRetrievedPerPrompt.ValueInt64()
+		}
+		if IsKnown(data.VectorDatabaseSettings.MaxTokens) {
+			createLLMBlueprintRequest.VectorDatabaseSettings.MaxTokens = data.VectorDatabaseSettings.MaxTokens.ValueInt64()
+		}
+	}
+
+	traceAPICall("CreateLLMBlueprint")
+	createResp, err := r.provider.service.CreateLLMBlueprint(ctx, createLLMBlueprintRequest)
 	if err != nil {
 		resp.Diagnostics.AddError("Error creating LLM Blueprint", err.Error())
 		return
@@ -210,6 +280,7 @@ func (r *LLMBlueprintResource) Read(ctx context.Context, req resource.ReadReques
 	if llmBlueprint.VectorDatabaseID != "" {
 		data.VectorDatabaseID = types.StringValue(llmBlueprint.VectorDatabaseID)
 	}
+	data.PromptType = types.StringValue(llmBlueprint.PromptType)
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
@@ -226,10 +297,8 @@ func (r *LLMBlueprintResource) Update(ctx context.Context, req resource.UpdateRe
 	_, err := r.provider.service.UpdateLLMBlueprint(ctx,
 		data.ID.ValueString(),
 		&client.UpdateLLMBlueprintRequest{
-			Name:             data.Name.ValueString(),
-			Description:      data.Description.ValueString(),
-			VectorDatabaseID: data.VectorDatabaseID.ValueString(),
-			LLMID:            data.LLMID.ValueString(),
+			Name:        data.Name.ValueString(),
+			Description: data.Description.ValueString(),
 		})
 	if err != nil {
 		if errors.Is(err, &client.NotFoundError{}) {