+++ pkg/provider/utils.go
@@ -27,6 +27,9 @@ import (
 
 const (
 	defaultTimeoutMinutes = 30
+
+	faithfulnessOpenAiRuntimeParam      = "MODERATION_OOTB_RESPONSE_FAITHFULNESS_OPENAI_API_KEY"
+	faithfulnessAzureOpenAiRuntimeParam = "MODERATION_OOTB_RESPONSE_FAITHFULNESS_AZURE_OPENAI_API_KEY"
 )
 
 type Knowable interface {
@@ -261,6 +264,11 @@ func formatRuntimeParameterValues(
 			continue
 		}
 
+		if isManagedByGuards(param) {
+			// skip the parameter if it is managed by guards
+			continue
+		}
+
 		parameter := RuntimeParameterValue{
 			Key:   types.StringValue(param.FieldName),
 			Type:  types.StringValue(param.Type),
@@ -285,6 +293,11 @@ func formatRuntimeParameterValues(
 	return listValueFromRuntimParameters(ctx, parameters)
 }
 
+func isManagedByGuards(param client.RuntimeParameter) bool {
+	return param.FieldName == faithfulnessOpenAiRuntimeParam ||
+		param.FieldName == faithfulnessAzureOpenAiRuntimeParam
+}
+
 func formatRuntimeParameterValue(paramType, paramValue string) (any, error) {
 	switch paramType {
 	case "boolean":
@@ -307,8 +320,55 @@ func listValueFromRuntimParameters(ctx context.Context, runtimeParameterValues [
 		}, runtimeParameterValues)
 }
 
+func prepareLocalFiles(folderPath types.String, files types.Dynamic) (localFiles []client.FileInfo, err error) {
+	localFiles = make([]client.FileInfo, 0)
+
+	if IsKnown(folderPath) {
+		folder := folderPath.ValueString()
+		if err = filepath.Walk(folder, func(path string, info os.FileInfo, innerErr error) error {
+			if innerErr != nil {
+				return innerErr
+			}
+			if info.IsDir() {
+				return nil
+			}
+
+			fileInfo, innerErr := getFileInfo(path, path)
+			if innerErr != nil {
+				return innerErr
+			}
+			localFiles = append(localFiles, fileInfo)
+
+			return nil
+		}); err != nil {
+			return
+		}
+	}
+
+	if IsKnown(files) {
+		var fileTuples []FileTuple
+		fileTuples, err = formatFiles(files)
+		if err != nil {
+			return
+		}
+
+		for _, file := range fileTuples {
+			var fileInfo client.FileInfo
+			fileInfo, err = getFileInfo(file.LocalPath, file.PathInModel)
+			if err != nil {
+				return
+			}
+
+			localFiles = append(localFiles, fileInfo)
+		}
+	}
+
+	return
+}
+
 func formatFiles(files types.Dynamic) ([]FileTuple, error) {
 	fileTuples := make([]FileTuple, 0)
+
 	switch value := files.UnderlyingValue().(type) {
 	case types.Tuple:
 		if len(value.Elements()) == 0 {
@@ -324,21 +384,27 @@ func formatFiles(files types.Dynamic) ([]FileTuple, error) {
 					PathInModel: filepath.Base(filePath),
 				})
 			case types.Tuple:
-				if len(v.Elements()) > 2 {
-					return nil, fmt.Errorf("files[%d] has more than 2 elements", i)
+				if len(v.Elements()) < 1 || len(v.Elements()) > 2 {
+					return nil, fmt.Errorf("files[%d] must have 1 or 2 elements", i)
 				}
 
-				localPath, isString1 := v.Elements()[0].(types.String)
-				pathInModel, isString2 := v.Elements()[1].(types.String)
-
-				if isString1 && isString2 {
-					fileTuples = append(fileTuples, FileTuple{
-						LocalPath:   localPath.ValueString(),
-						PathInModel: pathInModel.ValueString(),
-					})
-				} else {
+				localPath, ok := v.Elements()[0].(types.String)
+				if !ok {
 					return nil, fmt.Errorf("files[%d] has element that is not a string", i)
 				}
+				pathInModel := filepath.Base(localPath.ValueString())
+				if len(v.Elements()) == 2 {
+					modelPath, ok := v.Elements()[1].(types.String)
+					if !ok {
+						return nil, fmt.Errorf("files[%d] has element that is not a string", i)
+					}
+					pathInModel = modelPath.ValueString()
+				}
+
+				fileTuples = append(fileTuples, FileTuple{
+					LocalPath:   localPath.ValueString(),
+					PathInModel: pathInModel,
+				})
 			default:
 				return nil, errors.New("files must be a tuple of strings or tuples")
 			}
@@ -351,8 +417,6 @@ func formatFiles(files types.Dynamic) ([]FileTuple, error) {
 }
 
 func getFileInfo(localPath, pathInModel string) (fileInfo client.FileInfo, err error) {
-	fmt.Println("Reading file: ", localPath)
-
 	var fileReader *os.File
 	fileReader, err = os.Open(localPath)
 	if err != nil {